+++ sentry-ruby/spec/sentry/transaction_spec.rb
@@ -391,6 +391,16 @@ RSpec.describe Sentry::Transaction do
       expect(event[:transaction]).to eq("foo")
     end
 
+    it "finishes the transaction with explicit timestamp" do
+      timestamp = Sentry.utc_now.to_f
+      subject.finish(end_timestamp: timestamp)
+
+      expect(events.count).to eq(1)
+      event = events.last.to_hash
+
+      expect(event[:timestamp]).to eq(timestamp)
+    end
+
     it "assigns the transaction's tags" do
       Sentry.set_tags(name: "apple")
 
@@ -547,6 +557,21 @@ RSpec.describe Sentry::Transaction do
           "sample_rate" => "1.0"
         })
       end
+    end
+  end
+
+  describe "#set_name" do
+    it "sets name and source directly" do
+      subject.set_name("bar", source: :url)
+      expect(subject.name).to eq("bar")
+      expect(subject.source).to eq(:url)
+    end
+  end
+
+  describe "#set_context" do
+    it "sets arbitrary context" do
+      subject.set_context(:foo, { bar: 42 })
+      expect(subject.contexts).to eq({ foo: { bar: 42 } })
     end
   end
 end